Backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
dj-static==0.0.6
django==4.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.1.0
gunicorn==20.1.0
idna==3.3
mongoengine==0.23.1
pymongo==3.12.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
